+++ src/tsp.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 
-unsigned int tspBT(const float **dists, unsigned int n, unsigned int path[]);
+float tspBT(const float **dists, unsigned int n, unsigned int path[]);
 void tspBTRec(const float **dists, unsigned int n, unsigned int curIndex, float curDist, unsigned int curPath[], float &minDist, unsigned int path[]);
 
 #endif
